raw/pyproject.toml
[tool.poetry]
name = "raw_microsoft_wwi"
version = "1.4.0"
description = "Raw WWI Microsoft loader"
authors = []
readme = 'README.md'

[tool.poetry.scripts]
console = "run_locally:cli"

[tool.poetry.dependencies]
python = "^3.8"
delta-spark = "^1.2.1"
python-dotenv = "^0.20.0"
requests = "^2.28.1"


[tool.poetry.dev-dependencies]
click = "^8.1.3"
arrow = "^1.2.2"
flake8 = "^5.0.4"


[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
